\name{Lsdv.formula}
\alias{Lsdv.formula}
\title{formula}
\usage{
\method{Lsdv}{formula}(formula, data = list(), n, t, ...)
}
\arguments{
  \item{formula}{PIB~INF+TIR}

  \item{data}{the dataframe}

  \item{n}{the number of section}

  \item{t}{the time per section}

  \item{...}{not used}
}
\description{
formula
}

